refactor(client): use the new discography data and restructure some file structure
